Changed all cooldown times from ints to floats. Added tower booster super power.

+++ core/src/main/buildings/Building.java
@@ -10,7 +10,7 @@ public abstract class Building extends BoardObject {
 	
 
 
-	protected Building(String name, int x, int y, float size, float radius, int cooldown, int cost){
+	protected Building(String name, int x, int y, float size, float radius, float cooldown, int cost){
 		super(name, x, y, size, cost);
 		this.radius = radius;
 		this.cooldown = new CooldownObject(cooldown);
