+++ core/src/main/towers/Minutemen.java
@@ -11,7 +11,7 @@ public class Minutemen extends Tower {
     private static final int RADIUS = 1000;
     private static final String NAME = "MINUTEMEN";
     private static final int COST = 50;
-    private static final int COOLDOWN = 50;
+    private static final float COOLDOWN = 50;
     private static String DESCRIPTION = "Temporary soldiers with high range but low rate of fire";
     private static final float SIZE = 50;
     private static final float DAMAGE = 50;
