+++ core/src/main/towers/Sniper.java
@@ -13,7 +13,7 @@ public class Sniper extends Tower {
     private static final int RADIUS = 1000;
     private static final String NAME = "SNIPER";
     private static final int COST = 150;
-    private static final int COOLDOWN = 75;
+    private static final float COOLDOWN = 75;
     private static String DESCRIPTION = "Very long range";
     private static final float SIZE = 50;
     private static final float DAMAGE = 40;
